+++ pages/dashboard/addart.js
@@ -44,7 +44,7 @@ function AddArt(props) {
     form.imageUrl.forEach((file) => sendForm.append("imageUrl", file));
     try {
       const res = await axios
-        .post(`${process.env.NEXT_PUBLIC_SERVER}/api/auth/art`, sendForm, {
+        .post(`${process.env.NEXT_PUBLIC_SERVER}/api/art`, sendForm, {
           headers: { "content-type": "multipart/form-data" },
         })
         .then((res) => {
